docs: add project structure
