+++ app/(tabs)/home.tsx
@@ -28,8 +28,8 @@ import FixedHeaderOverlay from "../../components/FixedHeaderOverlay";
 import FilterHeader from "../../components/FilterHeader";
 import { MarketplaceItem, images } from "../../components/types";
 import { supabase } from "../../src/utils/supabase";
-import { supabaseApi } from "../../lib/supabase";
-import { setupSupabase, getMockMarketplaces } from "../../lib/supabaseSetup";
+import { supabaseApi } from "../src/lib/supabase";
+import { setupSupabase, getMockMarketplaces } from "../src/lib/supabaseSetup";
 import { useFilters } from "../../src/context/FilterContext";
 
 const { width, height } = Dimensions.get("window");
