+++ src/components/FixedHeaderOverlay.tsx
