+++ src/components/ideaHeader.tsx
